+++ src/components/PageNav.js
@@ -1,29 +1,33 @@
 import React from "react";
-import { Link, useLocation } from "react-router-dom"; 
+import { Link, useLocation } from "react-router-dom";
 import styles from "./PageNav.module.css";
 
 // 리소스
 import pandaImg from "../images/panda_img.svg";
 import logoText from "../images/logo_text.svg";
 
-
 //컴포넌트
-import LinkBtn from "./common/LinkBtn";
 import TextBtn from "./common/TextBtn";
 
 function PageNavRender() {
-  const location = useLocation(); 
+  const location = useLocation();
 
   const BoardList = () => {
     return (
       <ul className={styles.boardList}>
         <li>
-          <Link to="/free" className={location.pathname === "/free" ? styles.boardActive : ""}>
+          <Link
+            to="/free"
+            className={location.pathname === "/free" ? styles.boardActive : ""}
+          >
             자유게시판
           </Link>
         </li>
         <li>
-          <Link to="/items" className={location.pathname === "/items" ? styles.boardActive : ""}>
+          <Link
+            to="/items"
+            className={location.pathname === "/items" ? styles.boardActive : ""}
+          >
             중고마켓
           </Link>
         </li>
@@ -33,7 +37,7 @@ function PageNavRender() {
 
   return (
     <nav className={styles.outlineContain}>
-      <section className={styles.pageNavBox}>
+      <section className={styles.pageNavSection}>
         <Link to="/">
           <div className={styles.logoBox}>
             <img className={styles.logoImg} src={pandaImg} alt="logo_img" />
@@ -44,8 +48,6 @@ function PageNavRender() {
         <TextBtn text={"로그인"} />
       </section>
     </nav>
-
-
   );
 }
